requirements.txt
python-telegram-bot==13.3
aiohttp==3.7.3
python-socketio[client]==4.6.1
requests==2.24.0
mongoengine==0.22.1
pymongo==3.11.2
